some new benchmarks

+++ Testgen/Testgen_Sonam/Testgen/src/src/symbolicExec.c
@@ -212,7 +212,7 @@ void updateValBySymbolicName(char *sname, void *value) {
       else
        { 
           if((*(int*)value) < 0)
-          { updateFloatValBySname(sname, (*(int *)value));}
+          { updateFloatValBySname(sname, (*(float *)value));}
          else
           updateFloatValBySname(sname, (*(float *)value));}
       return;
@@ -238,7 +238,7 @@ void updateValBySymbolicName(char *sname, void *value) {
    {  if((*(int*)value) < 0)
           { updateFloatValBySname(sname, (*(int *)value));}
      else
-     updateFloatValBySname(sname, (*(int *)value));
+     updateFloatValBySname(sname, (*(float *)value));
    }
 }
